backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
mozilla-django-oidc>=5.0,<6.0
PyJWT>=2.8,<3.0
requests>=2.31,<3.0
cryptography>=42.0
